--- public/data/Data.js
@@ -1,13 +1,0 @@
-export const ASIDE_LIST = [
-  { id: 1, title: '소개' },
-  { id: 2, title: '도움말' },
-  { id: 3, title: '홍보 센터' },
-  { id: 4, title: 'API' },
-  { id: 5, title: '채용 정보' },
-  { id: 6, title: '개인정보처리방침' },
-  { id: 7, title: '약관' },
-  { id: 8, title: '위치' },
-  { id: 9, title: '인기 계정' },
-  { id: 10, title: '해시태그' },
-  { id: 11, title: '언어' },
-];
