+++ src/pages/jinkyung/Main/Feeds/Feeds.scss
@@ -5,102 +5,7 @@
   position: absolute;
   left: 17%;
   width: 45%;
-  height: 100vh;
+  height: 700px;
   padding: 40px 10px 0px;
-}
-
-/* -----article----- */
-#article {
-  @include background-color-white;
-  @include border;
-
-  /* -----user----- */
-  #user_wrap {
-    @include background-color-white;
-    display: flex;
-    justify-content: space-between;
-    height: 50px;
-
-    #user_info {
-      display: flex;
-      padding: 10px;
-
-      .user_img {
-        width: 35px;
-        height: 30px;
-        margin-right: 10px;
-        border-radius: 50%;
-        background-image: url(../../../../assets/images/jinkyung/Main/user_img.JPG);
-        background-size: cover;
-      }
-
-      .user_name {
-        width: 70%;
-        line-height: 30px;
-      }
-    }
-
-    #more {
-      width: 20px;
-      height: 20px;
-      margin: 15px 5px 0 0;
-      background-image: url(../../../../assets/images/jinkyung/Main/ellipsis.png);
-      background-size: cover;
-      cursor: pointer;
-    }
-  }
-
-  /* -----content img----- */
-  #content_img {
-    aspect-ratio: 1/1;
-    background-image: url(../../../../assets/images/jinkyung/Main/content_img.JPG);
-    background-size: cover;
-  }
-
-  /* -----react----- */
-  #react_wrap {
-    display: flex;
-    justify-content: space-between;
-    height: 40px;
-
-    #react_left {
-      display: flex;
-      padding: 8px 10px;
-
-      & > div {
-        @include icon;
-      }
-
-      /* stylelint-disable-next-line selector-class-pattern */
-      .icon_1 {
-        background-image: url(../../../../assets/images/jinkyung/Main/empty_heart.png);
-        cursor: pointer;
-
-        &.active {
-          background-image: url(../../../../assets/images/jinkyung/Main/heart.png);
-        }
-      }
-
-      /* stylelint-disable-next-line selector-class-pattern */
-      .icon_2 {
-        background-image: url(../../../../assets/images/jinkyung/Main/chat-balloon.png);
-      }
-
-      /* stylelint-disable-next-line selector-class-pattern */
-      .icon_3 {
-        background-image: url(../../../../assets/images/jinkyung/Main/upload.png);
-      }
-    }
-
-    #bookmark_wrap {
-      width: 50px;
-      height: 100%;
-      padding: 8px 11px;
-
-      .bookmark {
-        @include icon;
-        background-image: url(../../../../assets/images/jinkyung/Main/ribbon.png);
-      }
-    }
-  }
+  overflow: scroll;
 }
