+++ src/pages/jinkyung/Main/Feeds/ReactCount/ReactCount.scss
@@ -11,8 +11,6 @@
     height: 30px;
     margin-right: 10px;
     border-radius: 50%;
-    background-image: url(../../../../../assets/images/jinkyung/Main/comment_user_img.JPG);
-    background-size: cover;
   }
 
   /* stylelint-disable-next-line selector-class-pattern */
